# Only concatenate after all batches are done