Based on the discussion in #398, I was able to make `Llama.eval` faster by only calling `np.concatenate` once all the batches are done.

Performance measurements based on the [tester](https://github.com/abetlen/llama-cpp-python/issues/398#issue-1762827269) included in the original issue.

Master: (~3137.75ms overhead)
```py
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
      501    2.977    0.006   87.771    0.175 llama.py:400(eval)
```

This PR: (~350.9ms overhead)
```py
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
      501    0.041    0.000   84.657    0.169 llama.py:400(eval)

```