@shouyiwang 2023-06-26
It works!

Later on, I discovered that when I used my original test scripts and opted for the 7b model instead of the 33b model, and increased the length of the generation to a larger number such as 1500, the discrepancy became much more pronounced. And please note that the discrepancy only becomes apparent if your GPU is powerful enough. For example, if I limit the power of my 4090 to 150W, the speed difference between llama.cpp and llama-cpp-python is very small.

I tried testing the llama.cpp, the master branch and this PR on a 4090, and here are the results:

llama.cpp:
```
llama_print_timings:      sample time =   622.44 ms /  1500 runs   (    0.41 ms per token,  2409.87 tokens per second)
llama_print_timings: prompt eval time =   686.74 ms /   760 tokens (    0.90 ms per token,  1106.68 tokens per second)
llama_print_timings:        eval time = 13915.99 ms /  1498 runs   (    9.29 ms per token,   107.65 tokens per second)
llama_print_timings:       total time = 15485.60 ms
```

Master branch:
```
llama_print_timings:      sample time =   624.38 ms /  1499 runs   (    0.42 ms per token,  2400.77 tokens per second)
llama_print_timings: prompt eval time =    81.84 ms /     9 tokens (    9.09 ms per token,   109.96 tokens per second)
llama_print_timings:        eval time = 14081.58 ms /  1498 runs   (    9.40 ms per token,   106.38 tokens per second)
llama_print_timings:       total time = 31905.15 ms
```

This PR:
```
llama_print_timings:      sample time =   622.46 ms /  1499 runs   (    0.42 ms per token,  2408.17 tokens per second)
llama_print_timings: prompt eval time =    82.13 ms /     9 tokens (    9.13 ms per token,   109.59 tokens per second)
llama_print_timings:        eval time = 13908.38 ms /  1498 runs   (    9.28 ms per token,   107.70 tokens per second)
llama_print_timings:       total time = 16918.75 ms
```

This PR improved the total time from 31.9s to 16.9s. Well done!
Thank you so much for greatly improving the speed!