@shouyiwang 2023-06-26
GPU (4090) utilization every 0.5 seconds for the above test:

llama.cpp:
`0%, 0%, 2%, 20%, 51%, 87%, 86%, 87%, 87%, 87%, 87%, 87%, 88%, 88%, 87%, 88%, 88%, 88%, 88%, 88%, 88%, 88%, 89%, 89%, 89%, 89%, 89%, 89%, 89%, 89%, 98%, 48%, 0%, 0%, 0%, 0%,`

Main branch: 
`0%, 0%, 0%, 5%, 56%, 12%, 74%, 69%, 64%, 63%, 60%, 56%, 56%, 53%, 50%, 49%, 47%, 49%, 46%, 45%, 46%, 43%, 42%, 40%, 40%, 42%, 40%, 40%, 39%, 38%, 36%, 37%, 36%, 38%, 37%, 37%, 34%, 33%, 37%, 33%, 33%, 33%, 35%, 33%, 33%, 33%, 32%, 34%, 31%, 29%, 30%, 34%, 30%, 34%, 32%, 31%, 34%, 30%, 31%, 30%, 33%, 0%, 0%,`

This PR:
`0%, 0%, 5%, 49%, 44%, 78%, 77%, 78%, 78%, 78%, 77%, 77%, 77%, 78%, 77%, 77%, 77%, 77%, 78%, 78%, 76%, 76%, 77%, 77%, 76%, 76%, 76%, 77%, 77%, 77%, 77%, 76%, 76%, 45%, 0%, 0%,`

This PR doesn't completely solve the problem, but it's a significant improvement.