@abetlen 2023-06-26
@samfundev great catch, I'll go through profiling the eval again this week!